Add initial tide deployment.
The first draft here only contains logic for serial merges, batches
aren't implemented yet. There are a handful of TODOs in various
places, and there is only test coverage for the dividePool method.

Importantly, it does not actually merge or trigger tests, but instead
it simply logs out what actions it would take.