Add a prow plugin to label WIP pull requests
Developers often add prefixes to their pull requests to indicate that
the implementation is still a work in progress. A new WIP plugin will
use that information to add a label to these pull requests that will be
honored by the submit-queue. When the prefix is removed from the title
of the pull request, the plugin will remove the label and the submit
queue will consider the pull request for merging us usual.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>