Pin gcloud version to 163.0.0 for ci-kubernetes-e2e-gce-gci-qa-m60
K8s test team updated gcloud version on the testing machines. But
ci-kubernetes-e2e-gce-gci-qa-m60 still runs against the built-in
k8s version (v1.6.4) of COS m60, which is incompatible with the
updated gcloud version. This commit pins gcloud to the older
version 163.0.0 for ci-kubernetes-e2e-gce-gci-qa-m60.

This commit also delete tests for m59 as it has been deprecated.