Create a downward API struct for ProwJobs
In order to be more clear about the API that Prow will provide to the
agents that realize `ProwJobs`, the `pjutil.JobSpec` struct serializes
the full API and the `EnvForSpec()` utility now exposes all of the
fields in the struct as environment variables as well as exposing the
full JSON-encoded struct as `$JOB_SPEC`. Job code running on an agent
can opt to use either the JSON-encoded data or the more fine-grained
environment variables and can be certain they will never drift.

The `plank` controller will now provide the build identifier through the
`$BUILD_ID` environment variable to containers in `Pod`s in addition to
continuing to provide it as `$BUILD_NUMBER`. Job code that expects
`$BUILD_NUMBER` should migrate to the newer environment variable
whenever possible. In the future, client code should not make use of
`$BUILD_NUMBER` in any way as this is not a field that Prow can control
in Jenkins-agent jobs.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>