Ensure Jenkins jobs get a backwards compatible environment set
When migrating away from the old `$buildId` variable to the new
`$BUILD_ID` variable that both Jenkins and Kubernetes jobs can share, we
need to ensure the migration happens in a backwards compatible manner so
that dependent jobs can migrate at their own pace. This means passing
the old variable as well as the new one for now.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>