Make clonerefs best-effort on $JOB_SPEC, ignore periodics
The `clonerefs` utility can be used outside of a Prow context as long as
it is best-effort at honoring `$JOB_SPEC`. Furthermore, in periodic jobs
Prow is not delcarative as far as what Git refs are to be tested in the
job, so we should not attempt to clone anything from those types of job
configurations.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>