Refactor clonerefs to take args using JSON
When the `clonerefs` tool is being launched by a program like Prow, it
does not make sense to take rich, structured data and flatten it into
simple strings to pass arguments between the two programs. We can use a
JSON configuration file to pass the minimal set of configuration options
necessary for the tool to function. This option can coexist with the
current set of flags, which were written to be backwards compatible with
those on `bootstrap.py` and will be useful if a human is to use the
tool.

This patch also introduces a new field, `PathAlias`, into the `Refs`
structure. This is necessary to ensure that Prow is declarative in how
it instructs jobs to clone their refs. This field is currently only set
by the `clonerefs` tooling when a user passes the correct flags but we
will ask users to pass this information in through Prow job
configuration in the future.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>