prow: pod-utils: improve no-clone behavior
This patch adds the `skip_cloning` option to decorated jobs, which will
not add the `clonerefs` container even if there are refs to clone. Refs
will continue to be communiated in `$JOB_SPEC`, etc, so that the job can
do with them what it may.

Code volumes and working directories are also now not created on the Pod
when no cloning is occurring to clean up the `PodSpec`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>