prow: updateconfig: allow for custom keys in ConfigMaps
We now support updating specific keys in a `ConfigMap` and will do
incremental updates so as to not clobber data updated from other files.
This allows for one `ConfigMap` on the cluster to have data in its keys
updated from separate files on merge, as well as custom key values for
each file updated.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>