Allow job branch restrictions to be regex
When we have a branch `release-1.9` and want to quickly and quietly
create a derivative branch `release-1.9-hotfix-cve-1234`, we want the
derivative branch to have all of the same branch jobs as the branch it
is being derived from. In general, it should be cheap and easy to branch
out without having to modify too much of the testing infrastructure. One
step to making this possible is allowing branch restrictions on jobs to
be regular expressions. Git refs already have very stringent validation,
so we should not be concerned with trying to determine if a branch ref
as given is a regular expression or explicit branch name.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>