Allow authors and collaborators to close and reopen
Before this commit, closing are reopening required
actual assignment to the issue or the PR.

This meant that if you needed to reopen an issue, you would
need to use three commands - "/assign", "/reopen" and "/unassign".
This is not really helpful.

This commit allows only authors and collaborators to close and
reopen issues. It does not require assignment to the issue or PR.

A comment is also added before closing or reopening an issue or PR
to make sure that there is an audit trail for who asked to close or
reopen. This is necessary to make sure that there is no abuse of these
commands.