Don't exclude the `needs-ok-to-test` label for merge
If the lgtm plugin is enabled and is required for merge, the query
should not exclude the `needs-ok-to-test` label.

The lgtm plugin triggers one round of testing when applied to an
untrusted PR and removes the lgtm label if the PR changes so it
indicates to Tide that the current version of the PR is considered
trusted and can be retested safely.