Use generated Pod client from .kube/config in sinker
The custom `Cluster` file containing a subset of auth methods (only
certificate auth) is legacy and deprecated in this commit for `sinker`.
A proper `.kube/config` file should be used and all contexts present in
that file will be considered build cluster contexts that should be
reaped by the tool. The `Cluster` file will continue to be supported for
backwards compatibility until further notice.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>