Don't pass around possibly-nil configurations
The Trigger configuration is a well-formed Go struct that follows best
practices and is perfectly usable in the empty form. Every field in the
configuration must be set to enable some functionality, so the default
state of an uninitialized struct is the same as if a user did not have
any configuration at all.

Instead of passing around pointers and checking to see if the
configuration is missing, we should just use the default values as they
were intended.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>