deck: Display a heat map of the last 12 hours of job successes/failures
When triaging job failures it is common to want to identify correlations
in time, which the list of builds not well suited to. Use a heat map to
display the last 12 hours of the selected runs, newest to oldest, with the
y dimension being the duration of the build, the color of the square
being based on whether there were any failures, and the opacity of the
square increasing as the number of successes or failures increase.
If any job failed within a square, it is always red.

Each colored square is clickable which scrolls the page to the first
sample in that square.

The end goal is to make time-correlated failures across multiple jobs
more obvious. For instance, looking at a PR job shows a vertical line
as all the jobs kick off, while looking at a single job across many PRs
can show the exact time when test started regressing, and the user can
then scan into deeper detail.