Make Tide trigger batches before triggering serial tests.
This is important in order to ensure that the serially tested PR is a
member of the batch. This does not occur if we trigger the serial tests
first because that makes the PR pending and prevents it from being
included in the batch when we pick the batch members next sync loop.