Add pprof debug output to prow controllers
Having debug output available over HTTP allows for an admin to use
`kubectl port-forward` to grab data from the process on-demand. This is
incredibly useful to debug errant behavior and will enable us to respond
to resource usage worries. Go authors declare that they run with these
endpoints on at all times in production without seeing overhead for the
profiling:

https://groups.google.com/d/msg/golang-nuts/e6lB8ENbIw8/azeTCGj7AgAJ

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>