Provide a simple summary of success rate for the currently selected jobs
To assist visual estimation of the success rate for the set of selected
prow jobs, provide a small summary below the heat map that describes the
% of jobs that succeeded in the last 3, 12, and 48 hours, truncated to
whatever has been pruned from prow.

E.g.

    Success rate over time: 3h: 68%, 12h: 67%, 48: 74%

Which allows a human to leverage the job filter functions to quickly
answer assess a subset of jobs for trend (either by job name or by
repository).

The summary is displayed below the heat map to avoid cluttering other
higher visibility areas.