Add plugin to automatically apply milestone after PR merge
The milestoneapplier plugin configuration allows to define the current
milestone for each branch in a repo.
For example, `master: v1.16`, `release-1.15: v1.15`, etc.

For the default branch of the repo, generally `master`, it applies the
configured milestone after a PR is merged against the branch (even if
the PR had an older/different milestone).

For non-default branches, like `release-1.15`, it applies the configured
milestone after a (cherry-pick) PR is opened _and_ when the PR is
merged (even if the PR had an older/different milestone).

Note that it doesn't add the milestone when the PR is synced/updated
because we don't want the milestone to be added again if it's manually
removed.
Example scenario: if `v1.15.1` is the current milestone for
`release-1.15` branch, and a cherry-pick PR is opened but we want to
move it to `v1.15.2`, the milestone shouldn't be automatically updated
to `v1.15.1`.