Parse out JSONRPC errors, comment conditionally on update
Even when the JSONRPC server returns a 200, it can return errors in the
payload coming back to the client. We should parse those out and return
them as errors in the request.

Parsing out the JSONRPC response also allows us to conditionally comment
when we are updating the external bug trackers so we don't mislead users
into thinking we made changes that we did not.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>