Allow validation of dependent bugs in Bugzilla
Often one bug does not tell the full story and many bugs are linked
together with the "Depends On" field in Bugzilla. When considering a bug
valid, we should be able to say something about the states of dependent
bugs as it is often not appropriate to tackle a dependent child bug
before parent bugs have been processed.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>