Allow for registering cleanup work on interrupt
Often servers and other processes need to do custom cleanup work at
shutodwn that is not captured with the built-in cancellation functions.
The `OnInterrupt()` function allows for users to register this work in a
way that will be tracked, which will ensure that these processes are
given time to finish their cleanup before exiting.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>