prow/plugins/bugzilla: When dependent params are given, require a dependent bug
The OpenShift policy (reflected in the config in [1]) currently
requires release-4.2 pull requests to reference a bug targeting 4.2.z
(already works).  It also requires those bugs to depend on a VERIFIED+
4.3.0 bug to ensure that we don't regress by fixing something in 4.2.z
that is unfixed in 4.3.  Before this commit, the Bugzilla bot would
happily accept 4.2.z bugs with no dependencies.  With this commit,
those bugs will generate errors like:

  expected Bugzilla bug 123 to depend on a bug targeting the 4.3.0
  release and in one of the following states: VERIFIED,
  RELEASE_PENDING, CLOSED (ERRATA), but no dependents were found

[1]: https://github.com/openshift/release/blob/348c6a786f1d8b5bdc1f752580724a1c6309a066/core-services/prow/02_config/_plugins.yaml#L1836-L1879