Remove uniqueness checking on files in ConfigUpdater
`files` are keys from a map `updater.Maps` defined as
`map[string]ConfigMapSpec`.

Originally, it checked if a file show showed up more than once
in the config, which was not even possible when it has already
become a map (redundant but not a bug).

The PR [1] introduced `clusters` config-updater plugin and added
`updater.SetDefaults()` into function `validateConfigUpdater`.

By mistake, the recurrences of the key `file` were checked in the
inner loop. It became a bug. We remove the check because it should
not happen in reality.

[1]. https://github.com/kubernetes/test-infra/pull/15028